app/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.1
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.10.0
attrs==25.4.0
bcrypt==4.0.1
beautifulsoup4==4.13.5
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.116.1
fastapi-cli==0.0.11
fastapi-cloud-cli==0.1.5
frozenlist==1.8.0
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
jinja2==3.1.6
joblib==1.5.2
markdown-it-py==4.0.0
markupsafe==3.0.2
mdurl==0.1.2
motor==3.7.1
mpmath==1.3.0
multidict==6.7.0
numpy==2.3.3
packaging==25.0
pandas==2.3.2
passlib==1.7.4
propcache==0.4.1
pyasn1==0.6.1
pydantic==2.11.7
pydantic-core==2.33.2
pydantic-settings==2.11.0
pygments==2.19.2
pymongo==4.15.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
pytz==2025.2
pyyaml==6.0.2
requests==2.32.5
rich==14.1.0
rich-toolkit==0.15.1
rignore==0.6.4
rsa==4.9.1
scikit-learn==1.7.2
scipy==1.16.3
sentry-sdk==2.37.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
starlette==0.47.3
sympy==1.14.0
threadpoolctl==3.6.0
typer==0.17.4
typing-extensions==4.15.0
typing-inspection==0.4.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
yarl==1.22.0